+++ data_adapter/main.py
@@ -16,3 +16,6 @@ if __name__ == "__main__":
     command = sys.argv[1]
     if command == "get_process":
         get_process(sys.argv[2], sys.argv[3], sys.argv[4])
+    if command == "download":
+        collection_url = input("Enter collection URL: ")
+        download_collection(collection_url)